% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_params_evidence.R
\name{gen_params_evidence}
\alias{gen_params_evidence}
\title{Generate Test Parameters}
\usage{
gen_params_evidence(prior_name = c("Wishart", "BGL", "GHS", "G_Wishart"))
}
\arguments{
\item{prior_name}{The name of the prior for being tested with preexisting
test parameters, this is one of 'Wishart', 'BGL', 'GHS', 'G_Wishart'}
}
\value{
A list of matrices representing test parameters dependent on the 
prior specified in prior_name
}
\description{
Generates predetermined parameters for testing the functionality of
the graphical evidence method
}
\examples{
# Generate test parameter matrices for G-Wishart prior 
gen_params_evidence('G_Wishart')
}
